[Change] パッケー師をよりシンプルなものに変更
