+++ app/src/marvelous/client/discord/__init__.py
@@ -1,5 +1,5 @@
-from .message_gateway import message_gateway
+from .message import message_gateway
 from .client import bot
 from .start import start
-from .reaction_event import ReactionEvent
+from .reaction import ReactionEvent
 from .fetch_reaction_event import fetch_reaction_event
