+++ app/src/marvelous/client/discord/reaction.py
@@ -1,8 +1,17 @@
+from dataclasses import dataclass
 import discord
-from . import ReactionEvent
 from . import client
 from marvelous.helpers import first_match
 from typing import Optional
+
+
+@dataclass()
+class ReactionEvent:
+    sender: discord.Member
+    receiver: discord.Member
+    channel: discord.TextChannel
+    emoji: discord.PartialEmoji
+    reaction: Optional[discord.Reaction]
 
 
 def get_reaction_from_message(emoji: discord.PartialEmoji, message: discord.Message) -> Optional[discord.Reaction]:
