+++ app/src/marvelous/models/__init__.py
@@ -1,5 +1,8 @@
-from . import entities
+from . import booing
+from . import daily_bonus
 from . import errors
-from . import events
-from . import parts
-from . import usecases
+from . import marvelous
+from . import reaction
+from . import super_marvelous
+from . import survival_bonus
+from . import user
