--- app/src/marvelous/models/entities/__init__.py
@@ -1,1 +1,0 @@
-from .user import User
