--- app/src/marvelous/models/errors/__init__.py
@@ -1,4 +1,0 @@
-from .model_error import ModelError
-from .user_not_found_error import UserNotFoundError
-from .already_exist_error import AlreadyExistError
-from .self_user_action_error import SelfUserReactionError
