--- app/src/marvelous/models/errors/model_error.py
@@ -1,2 +1,0 @@
-class ModelError(Exception):
-    pass
