--- app/src/marvelous/models/events/__init__.py
@@ -1,1 +1,0 @@
-from .reaction import *
