--- app/src/marvelous/models/events/reaction/__init__.py
@@ -1,4 +1,0 @@
-from .reaction import Reaction
-from .super_marvelous import SuperMarvelousResult, SuperMarvelousSettings, SuperMarvelousReaction
-from .marvelous import MarvelousResult, MarvelousSettings, MarvelousReaction
-from .booing import BooingResult, BooingReaction, BooingSettings
