--- app/src/marvelous/models/events/reaction/reaction.py
@@ -1,9 +1,0 @@
-from ...entities import User
-
-
-class Reaction:
-    def send(self, sender: User, receiver: User):
-        pass
-
-    def cancel(self, sender: User, receiver: User):
-        pass
