--- app/src/marvelous/models/parts/__init__.py
@@ -1,1 +1,0 @@
-from .daily_bonus import DailyBonus
