--- app/src/marvelous/models/usecases/daily_steps.py
@@ -1,6 +1,0 @@
-import marvelous.data_store as data_store
-
-
-def reset_daily_steps() -> None:
-    """デイリーボーナスをリセットする"""
-    data_store.users.reset_daily_steps()
