+++ app/src/marvelous/models/user.py
@@ -1,10 +1,24 @@
+from __future__ import annotations
+from dataclasses import dataclass
+from .daily_bonus import DailyBonus
 import marvelous.data_store as data_store
-from ..entities import User
-from ..errors import UserNotFoundError, AlreadyExistError
+from marvelous.models.errors import UserNotFoundError, AlreadyExistError
 from logging import getLogger
+from typing import Iterable
 
 
 logger = getLogger(__name__)
+
+
+@dataclass()
+class User:
+    discord_id: int = None
+    display_name: str = None
+    marvelous_bonus: DailyBonus = None
+    booing_penalty: DailyBonus = None
+    super_marvelous_left: int = None
+    survival_bonus_given: bool = None
+    point: int = None
 
 
 def is_user_exist(discord_id: int) -> bool:
@@ -32,3 +46,9 @@ def update_name(discord_id: int, name: str) -> None:
     user: User = get_user(discord_id)
     user.display_name = name
     data_store.users.update_user(user)
+
+
+def get_ranking() -> Iterable[User]:
+    """ユーザーランキングを取得する"""
+    users: Iterable[User] = data_store.users.get_users_marvelous_point_ranking()
+    return users
