+++ app/tests/marvelous/usecases/test_send_reaction.py
@@ -1,7 +1,7 @@
 import pytest_mock
 import pytest
 from unittest.mock import MagicMock, call
-import marvelous.models.usecases.reaction as target_package
+import marvelous.models.services.reaction as target_package
 
 
 def test_check_self_user_failed(mocker: pytest_mock.MockerFixture):
